:truck:  MOVE: game.interface.ts를 game.ts 로 변경
